Makes clearer distintion between differential drive robots and holonomic robots.
